chore: rename constant file to matches
